spring-01: simplify app - replace reading and parse questions into QuestionServiceImpl
