+++ spring-01/src/main/java/ru/otus/spring/service/OutputService.java
@@ -2,5 +2,5 @@ package ru.otus.spring.service;
 
 public interface OutputService {
 
-    void outputString(String s);
+    void printLn(String s);
 }
