+++ spring-01/src/main/java/ru/otus/spring/service/OutputServiceStream.java
@@ -11,7 +11,7 @@ public class OutputServiceStream implements OutputService {
     }
 
     @Override
-    public void outputString(String inner) {
+    public void printLn(String inner) {
         outputStream.println(inner);
     }
 }
