+++ spring-01/src/main/java/ru/otus/spring/service/QuestionMapper.java
@@ -2,28 +2,9 @@ package ru.otus.spring.service;
 
 import ru.otus.spring.domain.Question;
 
-import java.util.List;
-import java.util.Map;
-
 import static java.lang.System.lineSeparator;
-import static ru.otus.spring.service.CsvReader.PARAM_NAME_ID;
-import static ru.otus.spring.service.CsvReader.PARAM_NAME_QUESTION;
-import static ru.otus.spring.service.CsvReader.PARAM_NAME_ANSWERS;
-import static ru.otus.spring.service.CsvReader.PARAM_NAME_RIGHT;
 
 public class QuestionMapper {
-    public Question toQuestion(Map<String, String> questionString) {
-        return Question.builder()
-                .id(Integer.parseInt(questionString.get(PARAM_NAME_ID)))
-                .text(questionString.get(PARAM_NAME_QUESTION))
-                .answers(parseAnswers(questionString.get(PARAM_NAME_ANSWERS)))
-                .rightAnswer(Integer.parseInt(questionString.get(PARAM_NAME_RIGHT)))
-                .build();
-    }
-
-    private List<String> parseAnswers(String inner) {
-        return List.of(inner.split(","));
-    }
 
     public String toString(Question question) {
         StringBuilder builder = new StringBuilder();
