--- spring-01/src/main/java/ru/otus/spring/service/Reader.java
@@ -1,8 +1,0 @@
-package ru.otus.spring.service;
-
-import java.util.List;
-import java.util.Map;
-
-public interface Reader {
-    List<Map<String, String>> readAllLines();
-}
